pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "query-dataset"
version = "0.2.0"
description = "Query DuckDB, MotherDuck, and Azure Data Explorer from one CLI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-kusto-data>=4.6.3",
    "duckdb>=1.4.1,<1.5",
]

[project.scripts]
query-dataset = "query_dataset.cli:main_entry"

[tool.hatch.build.targets.wheel]
packages = ["src/query_dataset"]
